New Working Swerve
In order to get the swerve drive to work we had to create a new class for each indiviual motor. I am currently looking into making one class for the motors and one for the swerve unit as a whole. I will then use this to make the code cleaner and less redundant.